+++ packages/vscode-ui5-language-assistant/src/constants.ts
@@ -1,7 +1,7 @@
 export const COMMAND_OPEN_DEMOKIT = "UI5LanguageAssistant.command.openDemokit";
 export const LOGGING_LEVEL_CONFIG_PROP = "UI5LanguageAssistant.logging.level";
 export const MANIFEST_SCHEMA = "manifest-schema";
-export const SCHEMA_URI =
-  "https://raw.githubusercontent.com/SAP/ui5-manifest/master/schema.json";
 export const SPLIT_ATTRIBUTE_ON_FORMAT =
   "UI5LanguageAssistant.SplitAttributesOnFormat";
+
+export const PACKAGE_NAME = "vscode-ui5-language-assistant";