+++ packages/vscode-ui5-language-assistant/src/extension.ts
@@ -41,6 +41,7 @@ import {
   bindingSemanticTokensProvider,
 } from "./binding-semantic-token-provider";
 import { OPEN_FRAMEWORK } from "@ui5-language-assistant/constant";
+import type { ManifestVersionChange } from "@ui5-language-assistant/context";
 
 type UI5Model = {
   url: string;
@@ -66,6 +67,11 @@ function init(context: ExtensionContext): void {
     client.onNotification(
       "UI5LanguageAssistant/ui5Model",
       async (model: UI5Model): Promise<void> => await updateCurrentModel(model)
+    );
+    client.onNotification(
+      "UI5LanguageAssistant/manifestVersionChanged",
+      (manifestChange: ManifestVersionChange) =>
+        getManifestSchemaProvider(context, manifestChange)
     );
   });
 }