+++ packages/vscode-ui5-language-assistant/src/logger.ts
@@ -1,0 +1,6 @@
+import { getLogger as logger, ILogger } from "@ui5-language-assistant/logger";
+import { PACKAGE_NAME } from "./constants";
+
+export const getLogger = (): ILogger => {
+  return logger(PACKAGE_NAME);
+};